Rename application localization file and update imports
